% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBoostedTrees-package.R
\docType{package}
\name{ggBoostedTrees-package}
\alias{ggBoostedTrees}
\alias{ggBoostedTrees-package}
\title{ggBoostedTrees: Visually Exploring Boosted Tree Models}
\description{
Graphic elements for exploring boosted tree models. Each figure is produced
in two steps: a \verb{gg_boost_*()} extractor pulls a tidy data frame with a
documented column contract out of a fitted model, and \code{autoplot()} renders
that data frame. Adding a new modelling backend means writing extractor
methods only.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ehrlinger/ggBoostedTrees}
  \item \url{https://ehrlinger.github.io/ggBoostedTrees}
  \item Report bugs at \url{https://github.com/ehrlinger/ggBoostedTrees/issues}
}

}
\author{
\strong{Maintainer}: John Ehrlinger \email{john.ehrlinger@gmail.com} (\href{https://orcid.org/0000-0002-5340-5154}{ORCID})

Authors:
\itemize{
  \item John Ehrlinger \email{john.ehrlinger@gmail.com} (\href{https://orcid.org/0000-0002-5340-5154}{ORCID})
}

}
\keyword{internal}
